+++ composer.json
@@ -124,7 +124,7 @@
         "psr-4": {
             "EonX\\EasyActivity\\Tests\\": "packages/EasyActivity/tests",
             "EonX\\EasyApiPlatform\\Tests\\": "packages/EasyApiPlatform/tests",
-            "EonX\\EasyApiPlatform\\Tests\\Stubs\\App\\": "packages/EasyApiPlatform/tests/Stubs/app/src",
+            "EonX\\EasyApiPlatform\\Tests\\Fixtures\\App\\": "packages/EasyApiPlatform/tests/Fixtures/app/src",
             "EonX\\EasyApiToken\\Tests\\": "packages/EasyApiToken/tests",
             "EonX\\EasyAsync\\Tests\\": "packages/EasyAsync/tests",
             "EonX\\EasyAwsCredentialsFinder\\Tests\\": "packages/EasyAwsCredentialsFinder/tests",
