+++ packages/EasyApiPlatform/tests/Fixtures/app/src/ApiResource/RelatedOwningDummy.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Tests\Stubs\App\ApiResource;
+namespace EonX\EasyApiPlatform\Tests\Fixtures\App\ApiResource;
 
 use ApiPlatform\Metadata\ApiResource;
 use Doctrine\DBAL\Types\Types;
@@ -10,7 +10,7 @@ use Doctrine\ORM\Mapping as ORM;
 
 #[ApiResource(types: ['https://schema.org/Product'])]
 #[ORM\Entity]
-class RelatedOwnedDummy
+class RelatedOwningDummy
 {
     #[ORM\Column(type: Types::INTEGER)]
     #[ORM\Id]
@@ -20,9 +20,8 @@ class RelatedOwnedDummy
     #[ORM\Column(type: Types::STRING, nullable: true)]
     private ?string $name;
 
-    #[ORM\OneToOne(inversedBy: 'relatedOwnedDummy', targetEntity: Dummy::class, cascade: ['persist'])]
-    #[ORM\JoinColumn(nullable: false)]
-    private Dummy $owningDummy;
+    #[ORM\OneToOne(mappedBy: 'relatedOwningDummy', targetEntity: Dummy::class, cascade: ['persist'])]
+    private ?Dummy $ownedDummy;
 
     public function getId(): int
     {
@@ -34,12 +33,9 @@ class RelatedOwnedDummy
         return $this->name;
     }
 
-    /**
-     * Get owning dummy.
-     */
-    public function getOwningDummy(): ?Dummy
+    public function getOwnedDummy(): ?Dummy
     {
-        return $this->owningDummy;
+        return $this->ownedDummy;
     }
 
     public function setId(int $id): void
@@ -52,8 +48,12 @@ class RelatedOwnedDummy
         $this->name = $name;
     }
 
-    public function setOwningDummy(Dummy $owningDummy): void
+    public function setOwnedDummy(?Dummy $ownedDummy): void
     {
-        $this->owningDummy = $owningDummy;
+        $this->ownedDummy = $ownedDummy;
+
+        if ($this->ownedDummy !== null && $this !== $this->ownedDummy->getRelatedOwningDummy()) {
+            $this->ownedDummy->setRelatedOwningDummy($this);
+        }
     }
 }
