+++ packages/EasyApiPlatform/tests/Fixtures/app/src/Serializer/NameConverter/CustomConverter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Tests\Stubs\App\Serializer\NameConverter;
+namespace EonX\EasyApiPlatform\Tests\Fixtures\App\Serializer\NameConverter;
 
 use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
